Fix toon gewicht in dropdown bij laden pagina

+++ Workout/HttpClients/GoogleClient.cs
@@ -58,7 +58,7 @@ public class GoogleClient(HttpClient client, IOptions<JsonSerializerOptions> jso
         }
     }
 
-    public Task SavePersoonlijkeGegevensLogAsync(double gewicht)
+    public Task SavePersoonlijkeGegevensLogAsync(decimal gewicht)
     {
         var fileName = $"{DateTime.Now:yyyy-MM-dd} persoonlijke info.json";
